Move 8-bit integer vector into primitives
