+++ library/Rattletrap/Json.hs
@@ -18,7 +18,7 @@ import Rattletrap.Header
 import Rattletrap.Initialization
 import Rattletrap.Int32
 import Rattletrap.Primitive.Int8
-import Rattletrap.Int8Vector
+import Rattletrap.Primitive.Int8Vector
 import Rattletrap.KeyFrame
 import Rattletrap.List
 import Rattletrap.Mark
